finish aufgabe 1

+++ templates/vector.hpp
@@ -31,7 +31,7 @@ public:
 	friend std::ostream& operator<<(std::ostream& os, const Vector& v)
 	{
 		os << "[";
-		for (int i = 0; i < v.size()-1; i++) {
+		for (size_t i = 0; i < v.size()-1; i++) {
 			os << v[i] << ", ";
 		}
 		os << v[v.size() - 1] << "]" ;
